+++ server/src/main/java/password/pwm/PwmConstants.java
@@ -157,6 +157,7 @@ public abstract class PwmConstants
     public static final String PARAM_POST_LOGIN_URL = "posturl";
     public static final String PARAM_FILE_UPLOAD = "fileUpload";
     public static final String PARAM_RECOVERY_OAUTH_RESULT = "roauthResults";
+    public static final String PARAM_RECOVERY_OAUTH_CANCEL = "cancelOAuth";
     public static final String PARAM_SIGNED_FORM = "signedForm";
     public static final String PARAM_USERKEY = "userKey";
     public static final String PARAM_METHOD_CHOICE = "methodChoice";